Try to preserve constraint diagnostics, but I'm missing some big pieces I think

+++ rscause/rscause_compiler/src/resolve_types.rs
@@ -434,7 +434,7 @@ pub enum ConstraintDiagnostic {
     Unknown,
     PendingInference,
     Resolver(Breadcrumbs, String),
-    Inferred(String, Vec<(usize, TypeConstraint, ConstraintDiagnostic)>),
+    Inferred(String, Vec<(u64, TypeConstraint, ConstraintDiagnostic)>),
 }
 
 pub struct ResolveTypesContext {
